Capture and forward side mouse buttons (BTN_SIDE and BTN_EXTRA)
Adds support for capturing and forwarding the side mouse buttons some mouses
have, typically used in web browsers for previous/next website in history.
(Those are presented by GDK as buttons 8 and 9.)

spice-vdagent requires a corresponding patch to handle the buttons.

At the spice and protocol level, the required changes were previously merged:
* https://gitlab.freedesktop.org/spice/spice/-/issues/45
* https://gitlab.freedesktop.org/spice/spice/-/merge_requests/140
* https://gitlab.freedesktop.org/spice/spice-protocol/-/merge_requests/22

This commit is a rebased and revised version of the patch by Kevin Pouget:
https://www.spinics.net/lists/spice-devel/msg42582.html
(I have confirmed with him that it's OK that I submit the updated patches.)

Co-Authored-By: Kevin Pouget <kpouget@redhat.com>
Co-Authored-By: Joan Bruguera <joanbrugueram@gmail.com>