spice-gtk-session: cache atoms
At the moment, spice-gtk only sends a grab message to the vdagent
based on the retrieved atoms.

With the upcoming changes, spice-gtk will have to know which
targets were advertised outside of clipboard_get_targets() callback.

We could use gtk_clipboard_wait_for_targets() or
gtk_clipboard_wait_is_*_available(), but the targets are not cached
by GTK+ on wayland for some reason. So let's cache them in spice-gtk
to avoid having to talk to the clipboard owner.

Signed-off-by: Jakub Janků <jjanku@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>