Fix parsing the usbids file
This fixes a regression from 032ca202f839fe1c49cddfd2b0459f9fecc23c86
where a check on whether the usbids file contained anything but the
wrong variable was checked. This is visible on windows when trying to
share a USB device as devices won't get named based on the usbids file.
Same thing on linux if we can't read the name from the sysfs.

`usbids_vendor_info` is only initialized to something not NULL after it
was checked to be different from NULL which is obviously wrong. The
patch wanted to check that the `usbids_vendor_count` was different from
0 instead.

Acked-by: Frediano Ziglio <freddy77@gmail.com>